committing testAll suite (error, working to fix) and ActorTest(complete)
