+++ PlantsVZombies/test/ca/carleton/pvz/test/ActorTest.java
@@ -2,33 +2,30 @@ package ca.carleton.pvz.test;
 
 import static org.junit.Assert.*;
 
-import org.junit.After;
-import org.junit.AfterClass;
 import org.junit.Before;
-import org.junit.BeforeClass;
 import org.junit.Test;
 
+
+import ca.carleton.pvz.actor.Actor;
+
 public class ActorTest {
-
-	@BeforeClass
-	public static void setUpBeforeClass() throws Exception {
-	}
-
-	@AfterClass
-	public static void tearDownAfterClass() throws Exception {
-	}
+	
+	private Actor testActor;
 
 	@Before
 	public void setUp() throws Exception {
-	}
-
-	@After
-	public void tearDown() throws Exception {
+		testActor = new Actor();
 	}
 
 	@Test
-	public void test() {
-		fail("Not yet implemented");
+	public void testActor() {
+		assertTrue(testActor.getSprite() == null);
 	}
+
+	
+// tearDown() is not necessary here, as garbage collection of objects
+// after the test class concludes. Other things that consume system
+// resources may need tearDown() however.
+
 
 }
